Reorganize API a Bit
